Move DOMPurify hooks to chats module
